Implement TimeMgr

+++ src/CommonLib/Times/Times.h
@@ -1,5 +1,5 @@
 /* Time.h */
 
-double get_wall_time();
-double get_cpu_time();
+long double get_wall_time();
+long double get_cpu_time();
